Refactored diffusion_test.cc
